pretty-printer for = that is python-repr()-like (var_export() is unacceptible)
Summary: repr()-like meaning it supports copying the output and pasting it to
get back the variable, for types and conditions where it is possible/practical.
(so not for objects or self-containing arrays, etc).

even when repasting is not possible, it is still visually consistent (in a
reasonbaly compact-vs-clear style) and it syntax errors in those cases, so
are any surprise case where you can't repaste becomes clear to the user sooner
rather than later.

also consistencized ___phpshell___ etc to ___phpsh___ etc.

this commit improves eletuchy's previous commit to use var_export(), which is
not acceptible because it silently nulls out various things (resources,
private object members), and hangs (well, fatals) on self-containing arrays.

we had to do this var_dump() parsing hack bc php, typically, has a bunch of
half-solutions for this kind of problem, no non-hacky combination of which is
really bearable, afaict.

Reviewed By: eletuchy, marcel, march, ccheever

Test Plan:
php> =4
4
php> ='hi'
"hi"
php> ="a\nb"
"a\nb"
php> ='a
 ... b'
"a\nb"
php> =id_get_conn(1, 'w')
<object #29 of type managed_connection> {
  "_conn:private" => <resource #2222 of type mysql link>,
  "_conn_acquire_function:private" => "_get_db_managed_cb",
  "_conn_acquire_args:private" => array(
    ::snip-for-public-repo::
    6 => true,
    7 => 2,
    8 => 1,
  ),
  ::etc::
}
php> =___phpsh___parse_dump_test()
10
php>

Revert Plan: ok

DiffCamp Revision: 36882