stop showing php command prompt on every load
Summary: i'm not sure if ishermandom meant to add this or not, i
don't remember it from the diffcamp diff.  anyway, we used to show
this and stopped because it's too long now.  if we want, it makes
more sense to have this as a command, like 's' -> 'show php command
line' or something.

i also think people have associated this change with phpsh having an
error; i think instead that there are actually probably instead ape
issues on some boxes that have nothing to do with this command showing
now.

Reviewed By: marcel

Test Plan: load phpsh

Revert Plan: ok