kill setuptools; make local installs work sanely
Summary: setuptools is too crazy, and was making local installs
hard as well as doing crazy things like creating the egg staging
directory as root in the local directory (e.g. see commit cdb00b25).
the general python consensus seems to be that setuptools is broken,
and distutils with distro-based package management is the one and
only True Way, at least for now.

of course this means that the readline and pysqlite dependencies
will not be auto-grabbed, README updated.

we still clobber previous rc files in the install.
(/etc/phpsh/rc.php is clobbered in a root/global install,
~/.phpsh/rc.php is clobbered in a non-root/local install)
this will probably be changed soon but is independent of this change,
and besides no one has complained about it yet..

Reviewed By: marcel

Test Plan:
- testing local install after killing all the files of existing global
install and vice versa
- testing local/global documentation with
php> d array_rand
after killing one or the other of /etc/phpsh/php_manual.db and
~/.phpsh/php_manual.db

Revert Plan: ok